Add opened_only matcher for security groups.
This matcher allows the user to make exclusivity statements about certain rules.
Prior, a user could only make statements about the existence of opened rules for
security groups, now they can state that not only is there an opened rule but
that that is the _only_ open rule for a given port, protocol, cidr.