Support for vpc filter in the route tables
This commit changes the base resource adding the ability to pass extra
parameteres to them in the constructor. These parameters could be used
to filter the resource adding more context or for any other purpose.

The scenario that I'm trying to address here is the same than the issue
number #223. In the deploy that I'm working now, I have multiple route
tables with the same belonging to different VPCs.