Bug fixes
Fixed typo in exception.
Fixed bug regarding invoking describe_time_to_live method from the client of Aws::DynamoDB::Table class. The current way (before the fix) this was processed was to ignoring the two extra parameters passed to method_missing method and also not passing the required parameter of the table name. This might not be the best way to fix this, it looks like more a workaround. It seems a bit unproductive to have to pass again the table name, since the client itself is instantiated from the Aws::DynamoDB::Table instance, so a change in the API to hide complexity (in AWS SDK, the table and it's corresponding class are two different entities) this would be interesting too.