Fixing NoMethodError for AWS IAM policies
See https://github.com/k1LoW/awspec/issues/421
Added new methods to Awspec::Type::IamPolicy in order to use delegation and check the existence of the policy as returned by the respective finder.
Created the exception Awspec::NoExistingResource to raise an error if the policy is not available.
Extended the unit tests to include those use cases.