Bug fixes for issue 383
https://github.com/k1LoW/awspec/issues/383

Changes to implement the usage of Awspec::NonExistingResource exception into Awspec::Type::Ec2.
A new stub was created, because the SDK will always return a Aws::EC2::Types::DescribeInstancesResult with an empty reservations attribute.
Changed the Awspec::Helper::Finder::Ec2 to handle expecific exceptions instead of all of them.
Added tests to the spec to validate Awspec::NonExistingResource is returned.