Refactoring to fix Rubocop issues
The class Awspec::Type::Ec2 code is too large (by Rubocop configurations). An attempt to move the EC2 status to a new module was attempted, but the documentation generator requires the definition of a Awspec::Type::Ec2::STATE constant, so the refactoring was not complete.
The Awspec::Helper::AllStates implementation suggests to aggregate all states definition on a single module, also using a lazy Enumerator to reduce memory usage. It might (or not) give an option enable reusing of state constants.