:sparkles: There should be only one instance of Config
Therefore make it a Singleton so there's no confusion and we can
configure it in many places but not need to pass it in throughout the
code base when it's needed.

I'm normally _not_ a fan of Singleton, but in this specific case I think
it makes sense.

See: https://ruby-doc.org/stdlib-2.6.5/libdoc/singleton/rdoc/Singleton.html

Laying groundwork for #477