:sparkles: Start using Config in the ClientWrap
Since the ClientWrap was never getting arguments for the backoff logic
from it's collaborator (Finder.rb), it maybe makes sense to have
ClientWrap use the Config class directly, especially if the
configurations are targeted at the ClientWrap itself.

One issue this brings up is testing becomes a bit harder now since
Config is a Singleton. For now we just reset it back to it's original
state, but we could mock it out or have a mechanism to reset the
defaults.

Laying groundwork for #477