Added ability to test AWS resource integration paths
In API gateway, you can change how a resource is integrated into the backend of a service.
Often someone would like to keep the original resource path but change the path to the backend (i.e. to keep backwards compatibility of APIs).
I added the ability to test the integration paths so that you can see the path that integrates with the backend.